% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_table_.R
\name{active_age_grp_}
\alias{active_age_grp_}
\title{Calculate Age Grouping}
\usage{
active_age_grp_(dbl)
}
\arguments{
\item{dbl}{A \code{double} vector of ages}
}
\value{
A \code{factor} of age groups
}
\description{
Calculate Age Grouping
}
\keyword{internal}
