% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_table_.R
\name{active_calc_}
\alias{active_calc_}
\title{Calculate Active Cases by A Categorical Variable}
\usage{
active_calc_(data, grp = c("age", "sex", "race", "ethnicity"))
}
\arguments{
\item{data}{Data from the associated \verb{active_calc_*()} function}

\item{grp}{The grouping structure to use; only single-variable grouping is
currently supported}
}
\value{
A \code{tibble} with columns \code{grp} (\code{fct}), \code{n} (\code{int}),
\code{percent} (\code{dbl}), and \code{rate} (\code{dbl})
}
\description{
Calculate Active Cases by A Categorical Variable
}
\keyword{internal}
