% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_table_.R
\name{active_collapse_age_}
\alias{active_collapse_age_}
\title{Collapse Age Groupings}
\usage{
active_collapse_age_(data)
}
\arguments{
\item{data}{Age-grouped data from
\code{\link[covidReport:active_join_pop_]{active_join_pop_()}}}
}
\value{
\code{data} with age groups collapsed
}
\description{
Collapse Age Groupings
}
\keyword{internal}
