% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_map_rate.R
\name{active_map_rate}
\alias{active_map_rate}
\title{Plot Active Case Rates by ZIP Code}
\usage{
active_map_rate(
  data = pos(process_inv(read_inv(date = date))),
  days = 14L,
  date = NULL
)
}
\arguments{
\item{data}{Prepped case data}

\item{days}{Number of active days to consider}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
A \code{ggplot} object
}
\description{
Plot Active Case Rates by ZIP Code
}
