% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_plot_.R
\name{active_plot_}
\alias{active_plot_}
\title{Plot Active Cases by a Categorical Variable}
\usage{
active_plot_(data, grp = c("age", "sex", "race", "ethnicity"), date = NULL)
}
\arguments{
\item{data}{Data from an \verb{active_calc_*()} function}

\item{grp}{The grouping variable}
}
\value{
A \code{ggplot}
}
\description{
Plot Active Cases by a Categorical Variable
}
\keyword{internal}
