% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_table_.R
\name{active_race_grp_}
\alias{active_race_grp_}
\title{Calculate Race Grouping}
\usage{
active_race_grp_(chr)
}
\arguments{
\item{dbl}{A \code{character} vector of race}
}
\value{
A \code{factor} of racial groups
}
\description{
Calculate Race Grouping
}
\keyword{internal}
