% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_table_.R
\name{active_table_}
\alias{active_table_}
\title{Create Table of Active Cases by A Categorical Variable}
\usage{
active_table_(data, grp_lbl)
}
\arguments{
\item{data}{Data from the associated \verb{active_calc_*()} function}

\item{grp_lbl}{Label for grouping variable}
}
\value{
A \code{flextable}
}
\description{
Create Table of Active Cases by A Categorical Variable
}
\keyword{internal}
