% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tbl.R
\name{as_tbl}
\alias{as_tbl}
\alias{as_tbl.gt_tbl}
\alias{as_tbl.flextable}
\alias{as_tbl.character}
\title{Convert \code{flextable}, \code{gt_tbl}, and HTML \code{character} Objects to a \code{tibble}}
\usage{
as_tbl(x)

\method{as_tbl}{gt_tbl}(x)

\method{as_tbl}{flextable}(x)

\method{as_tbl}{character}(x)
}
\arguments{
\item{x}{A \code{flextable}, \code{gt_tbl}, or \code{character} vector defining an HTML
table}
}
\value{
A \code{tibble}
}
\description{
\code{as_tbl()} first converts objects to HTML, then extracts the HTML table from
the result
}
\keyword{internal}
