% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_table_active.R
\name{case_calc_active}
\alias{case_calc_active}
\title{Calculate Active, Deceased, and Inactive/Recovered COVID-19 Cases}
\usage{
case_calc_active(data = pos(process_inv(read_inv(date = date))), date = NULL)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process-nbs]{pos(process_inv())}}}

\item{date}{Download date of the data; defaults to most recent}
}
\value{
A \code{tibble}
}
\description{
Calculate Active, Deceased, and Inactive/Recovered COVID-19 Cases
}
\keyword{internal}
