% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_table_confirmed_probable.R
\name{case_table_confirmed_probable}
\alias{case_table_confirmed_probable}
\title{Tabulate Confirmed & Probable Cases/Deaths}
\usage{
case_table_confirmed_probable(
  data = coviData::process_positive_people(date = date),
  date = NULL
)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process_positive_people]{process_positive_people()}}}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
A \code{flextable} object
}
\description{
Tabulate Confirmed & Probable Cases/Deaths
}
