% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_table_summary.R
\name{case_table_summary}
\alias{case_table_summary}
\title{Summarize Cases for Weekend Reporting}
\usage{
case_table_summary(date = NULL, row = c(1L, 2L))
}
\arguments{
\item{date}{Report date to summarize}

\item{row}{Row of summary table to output; either \code{1} or \code{2}}
}
\value{
A \code{gt_tbl}
}
\description{
Summary numbers for weekend reporting. This outputs the first and second
rows of the table separately (using the \code{row} parameter). The first row
contains \emph{Total Cases}, \emph{Total Tests}, and \emph{Active Cases}; the second row
contains \emph{New Cases}, \emph{Inactive Cases}, and \emph{Total Deaths}.
}
