% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pop.R
\name{count_pop}
\alias{count_pop}
\title{Summarize Population Demopgraphics}
\usage{
count_pop(cols = c("age", "sex", "race", "ethnicity"))
}
\arguments{
\item{cols}{\code{character}. The column(s) to summarize.}
}
\value{
A \code{tibble} with a column for each
\code{cols} (\code{int} of \code{cols = "age"}, \code{fct} otherwise) and corresponding
\code{population} (\code{int})
}
\description{
Summarize Population Demopgraphics
}
\keyword{internal}
