% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coviData.R
\name{coviData-helpers}
\alias{coviData-helpers}
\title{COVID-19 Data Munging Tools for Shelby County Health Department}
\description{
Many of the data munging tools used in the collection of covid* packages
for SCHD belong to the coviData package. For convenience, these are
re-exported here.
}
\keyword{internal}
