% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_race.R
\name{death_calc_race}
\alias{death_calc_race}
\title{Calculate Death Rates and Percentages by Race}
\usage{
death_calc_race(data = pos(process_inv(read_inv(date))), date = NULL)
}
\arguments{
\item{data}{NBS case data, as returned by
\code{\link[coviData:read-nbs]{pos(process_inv())}}}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
A \code{tibble}
}
\description{
Calculate Death Rates and Percentages by Race
}
\keyword{internal}
