% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_sex.R
\name{death_table_sex}
\alias{death_table_sex}
\title{Tabluate Deaths by Sex}
\usage{
death_table_sex(data = pos(process_inv(read_inv(date))), date = NULL)
}
\arguments{
\item{data}{NBS case data, as returned by
\code{\link[coviData:read-nbs]{pos(process_inv())}}}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
A \code{flextable}
}
\description{
Tabluate Deaths by Sex
}
