% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_table_total.R
\name{death_table_total}
\alias{death_table_total}
\title{Tabulated Death and Survival Outcomes from COVID-19}
\usage{
death_table_total(
  data = coviData::process_positive_people(date = date),
  date = NULL
)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process_positive_people]{process_positive_people()}}}

\item{date}{The download data of the data; defaults to most recent}
}
\value{
A \code{gt_tbl}
}
\description{
Tabulated Death and Survival Outcomes from COVID-19
}
