% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_calc_.R
\name{demog_calc_}
\alias{demog_calc_}
\title{Calculate Demographic Summary}
\usage{
demog_calc_(data, grp = c("age", "sex", "race", "ethnicity"), peds = FALSE)
}
\arguments{
\item{data}{Data from the associated \verb{*_calc_*()} function}

\item{grp}{The grouping structure to use; only single-variable grouping is
currently supported}
}
\value{
A \code{tibble} with columns \code{grp} (\code{fct}), \code{n} (\code{int}),
\code{percent} (\code{dbl}), and \code{rate} (\code{dbl})
}
\description{
Calculate Demographic Summary
}
\keyword{internal}
