% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_calc_.R
\name{demog_collapse_age_}
\alias{demog_collapse_age_}
\title{Collapse Age Groupings}
\usage{
demog_collapse_age_(data)
}
\arguments{
\item{data}{Data with \code{numeric} ages}
}
\value{
\code{data} with age groups collapsed
}
\description{
Collapse Age Groupings
}
\keyword{internal}
