% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_calc_.R
\name{demog_collapse_race_}
\alias{demog_collapse_race_}
\title{Collapse Race Groupings}
\usage{
demog_collapse_race_(data)
}
\arguments{
\item{data}{Data with \code{character} race variable}
}
\value{
A \code{tibble} with collapsed race variable
}
\description{
Collapse Race Groupings
}
\keyword{internal}
