% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_calc_.R
\name{demog_join_}
\alias{demog_join_}
\title{Join to Population Demographics}
\usage{
demog_join_(data, grp = c("age", "sex", "race", "ethnicity"), peds = FALSE)
}
\arguments{
\item{data}{Data summarized by demographic}

\item{grp}{Demographic group}
}
\value{
A \code{tibble} with column \code{n_pop}
}
\description{
Join to Population Demographics
}
\keyword{internal}
