% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_calc_.R
\name{demog_relevel_race}
\alias{demog_relevel_race}
\title{Relevel Race for Appropriate Ordering}
\usage{
demog_relevel_race(data)
}
\arguments{
\item{data}{Data from a \verb{_calc_race()} function}
}
\value{
\code{data} with \code{grp} levels (and rows) reordered
}
\description{
Relevel Race for Appropriate Ordering
}
\keyword{internal}
