% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_table_.R
\name{demog_table_}
\alias{demog_table_}
\title{Create Demographic Table}
\usage{
demog_table_(data, grp_lbl)
}
\arguments{
\item{data}{Data from the associated \verb{*_calc_*()} function}

\item{grp_lbl}{Label for grouping variable}
}
\value{
A \code{flextable}
}
\description{
Create Demographic Table
}
\keyword{internal}
