% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_active.R
\name{filter_active}
\alias{filter_active}
\title{Filter Case Data to Active Cases}
\usage{
filter_active(data, days = 14L, date = NULL)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process-nbs]{pos(process_inv())}}}

\item{days}{\code{numeric}. The number of days a case is considered active}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
The input \code{data} filtered to rows containing deaths
}
\description{
Filter Case Data to Active Cases
}
\keyword{internal}
