% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_deaths.R
\name{filter_deaths}
\alias{filter_deaths}
\title{Filter Case Data to COVID-19 Deaths}
\usage{
filter_deaths(data, col = "die_from_illness_ind", cols_check = "inv_death_dt")
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process_positive_people]{process_positive_people()}}}

\item{col}{\verb{<tidy-select>} A single, \code{character} indicator column for
COVID-19 deaths. Deaths should be labeled \code{"Y"}.}

\item{cols_check}{\verb{<tidy-select>} Columns to check for missing data. Rows
with missing data are assumed to be coded incorrectly and will not be
counted as a death, even if \code{col} is labeled \code{"Y"}.}
}
\value{
The input \code{data} filtered to rows containing deaths
}
\description{
Filter Case Data to COVID-19 Deaths
}
\keyword{internal}
