% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peds.R
\name{filter_peds}
\alias{filter_peds}
\title{Filter to Pediatric COVID-19 Cases}
\usage{
filter_peds(data)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process-nbs]{pos(process_inv())}}}
}
\value{
\code{data} filtered to cases less than 18
}
\description{
Filter to Pediatric COVID-19 Cases
}
