% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{fmt_covid_table}
\alias{fmt_covid_table}
\title{Standardize \code{flextable} Styling for COVID-19 Reports}
\usage{
fmt_covid_table(
  flextable,
  total = FALSE,
  align_label = c("left", "center", "right")
)
}
\arguments{
\item{flextable}{A \code{flextable} object}

\item{total}{Does the table contain a "Total" row at the bottom?}

\item{align_label}{Alignment for left-most column (labels, by convention)}
}
\value{
A \code{flextable} with modified styling
}
\description{
Standardize \code{flextable} Styling for COVID-19 Reports
}
\keyword{internal}
