% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{ft_as_html}
\alias{ft_as_html}
\title{Convert a \code{flextable} to an HTML String}
\usage{
ft_as_html(flextable)
}
\arguments{
\item{flextable}{A \code{flextable} object}
}
\value{
A \code{character} vector of length 1 containing HTML content from a
\code{flextable}
}
\description{
Convert a \code{flextable} to an HTML String
}
\keyword{internal}
