% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_timeseries.R
\name{gs_timeseries}
\alias{gs_timeseries}
\title{Create Timeseries for Google Sheets Data}
\usage{
gs_timeseries(
  inv = process_inv(read_inv(date = date)),
  delay = 5L,
  date = NULL
)
}
\arguments{
\item{inv}{Investigation data from
\code{\link[coviData:process_inv]{process_inv()}}}

\item{delay}{The number of days that reporting is delayed; will truncate
percent positive numbers at this date}

\item{date}{Download date of the data}

\item{pcr}{PCR test data from
\code{\link[coviData:process_inv]{process_pcr()}}}
}
\value{
A \code{tibble} with columns
\code{dt} (\code{Date}), \code{inv_pos} (\code{int}), \code{inv_neg} (\code{int}), \code{inv_pct_pos} (\code{dbl}),
\code{inv_pct_pos_avg} (\code{dbl}), \code{pcr_pos} (\code{int}), \code{pcr_neg} (\code{int}),
\code{deaths} (\code{int})
}
\description{
Create Timeseries for Google Sheets Data
}
\keyword{internal}
