% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{insp_archive_table}
\alias{insp_archive_table}
\title{Create and Manage Inspection Table Archive Files}
\usage{
insp_archive_table(
 
    table_path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/",
    paste0("insp_table_", Sys.Date()), ext = "html"),
 
    archive_dir = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/Archive/"),
  table_pattern = "insp_table_.*html",
  force = FALSE
)
}
\arguments{
\item{table_path}{The path to the table to archive}

\item{archive_dir}{The path to the archive directory}

\item{table_pattern}{The pattern to search for when removing old tables from
the directory in \code{table_path}}

\item{force}{Should any existing file in \code{archive_dir} be overwritten if
there is a conflict?}
}
\value{
\code{table_path} (invisibly)
}
\description{
\code{insp_archive_table()} creates a backup of the file at \code{table_path}
and ensures that only one file matching \code{table_pattern} is in the directory
specified in \code{table_path}. It also ensures that backups are kept to a
reasonable number in \code{archive_dir} (at least seven at all times, possibly
more if multiple backups per day were made in the last seven days).
}
\seealso{
Other functions in the inspections data pipeline (
\code{
  \link[covidReport:insp_load_data]{insp_load_data()}
  },
\code{
  \link[covidReport:insp_prep_data]{insp_prep_data()}
  },
\code{
  \link[covidReport:insp_create_table]{insp_create_table()}
  },
\code{
  \link[covidReport:insp_save_table]{insp_save_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:insp_table_pipeline]{insp_table_pipeline()}
  })
}
