% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{insp_create_table}
\alias{insp_create_table}
\title{Create HTML Table of COVID-19 Business Inspections Results}
\usage{
insp_create_table(.data)
}
\arguments{
\item{.data}{The prepared inspections data}
}
\value{
A \code{reactable} object
}
\description{
\code{insp_create_table()} takes prepared inspections data and creates an
HTML table for publishing on the web (using the \strong{DT} package). Data
should first be prepared using
\code{\link[covidReport:insp_prep_data]{insp_prep_data()}}.
}
\seealso{
Other functions in the inspections data pipeline (
\code{
  \link[covidReport:insp_load_data]{insp_load_data()}
  },
\code{
  \link[covidReport:insp_prep_data]{insp_prep_data()}
  },
\code{
  \link[covidReport:insp_save_table]{insp_save_table()}
  },
\code{
  \link[covidReport:insp_archive_table]{insp_archive_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:insp_table_pipeline]{insp_table_pipeline()}
  })
}
