% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table2.R
\name{insp_load_data}
\alias{insp_load_data}
\title{Load Inspections Data}
\usage{
insp_load_data(
  path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/",
    "Grand List of Inspections", ext = "xlsx"),
  check = TRUE
)
}
\arguments{
\item{path}{The path to the excel workbook containing inspections data}

\item{check}{Should column names be checked? See \strong{Details} for more
information on name checking.}
}
\value{
A \code{tibble} containing the inspections data
}
\description{
\code{insp_load_data()} loads the data from an excel workbook into the R
session. standardizes column names, and optionally checks that the column
names match expected values.
}
\details{
If \code{check = TRUE}, \code{insp_load_data()} expects at least five columns
giving information on the (last?) inspection visit date, the business name,
the business address, the number of total violations by the business, and the
closure date (if it exists). The expected values of the column names are
formulated as generally as possible to allow for human error. The columns are
matched as follows:

\itemize{
\item \strong{inspection visit date} must contain the words \emph{"date"}
and \emph{"visit"}
\item \strong{business name} must contain the words \emph{"name"} and
\emph{"business"}
\item \strong{business address} must contain the word \emph{"address"}
\item \strong{number of violations} must contain the word
\emph{"violations"}
\item \strong{closure date} must contain the string \emph{"clos"} and the
word \emph{"date"}
}
}
\seealso{
Other functions in the inspections data pipeline (
\code{\link[covidReport:insp_prep_data]{insp_prep_data()}},
\code{
  \link[covidReport:insp_create_table]{insp_create_table()}
  },
\code{\link[covidReport:insp_save_table]{insp_save_table()}},
\code{
  \link[covidReport:insp_archive_table]{insp_archive_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:insp_table_pipeline]{insp_table_pipeline()}
  })
}
