% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{insp_save_table}
\alias{insp_save_table}
\title{Save an HTML Table Produced by \code{insp_create_table()}}
\usage{
insp_save_table(
  .table,
  path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/",
    paste0("insp_table_", Sys.Date()), ext = "html"),
  force = FALSE
)
}
\arguments{
\item{.table}{A \code{datatables} object}

\item{path}{The location to save the table}

\item{force}{Should an existing file at \code{path} be overwritten if a
conflict occurs?}
}
\value{
The input \code{.table} (invisibly)
}
\description{
\code{insp_save_table()} saves a \code{reactable} object to the specified
\code{path}. The defaults are intended for saving the inspections data table.
}
\details{
This function replies on \code{\link[htmlwidgets:saveWidget]{saveWidget()}},
but replaces the base64 encoded dependencies in the self-contained file with
the human-readable equivalent.
}
\seealso{
Other functions in the inspections data pipeline (
\code{
  \link[covidReport:insp_load_data]{insp_load_data()}
  },
\code{
  \link[covidReport:insp_prep_data]{insp_prep_data()}
  },
\code{
  \link[covidReport:insp_create_table]{insp_create_table()}
  },
\code{
  \link[covidReport:insp_archive_table]{insp_archive_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:insp_table_pipeline]{insp_table_pipeline()}
  })
}
