% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table2.R
\name{insp_trim_archive}
\alias{insp_trim_archive}
\title{Trim Archive Directory for Inspections Tables}
\usage{
insp_trim_archive(
 
    archive_dir = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/Archive/"),
  min_backups = 7L
)
}
\arguments{
\item{archive_dir}{The directory containing archived inspections tables}

\item{min_backups}{The minimum number of backups to retain}
}
\value{
\code{archive_dir} (invisibly)
}
\description{
\code{insp_trim_archive()} ensures that at least \code{min_backups} backups are
always kept of the inspections table, but that at most \code{min_backups} days
worth of backups are kept if there are more than \code{min_backups} backups. This
function is for internal use in
\code{
\link[covidReport:insp_archive_table]{insp_archive_table()}
}.
}
\seealso{
\code{
  \link[covidReport:insp_archive_table]{insp_archive_table()}
  }, \code{\link[coviData:trim_backups]{trim_backups()}}
}
\keyword{internal}
