% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{insp_trim_tables}
\alias{insp_trim_tables}
\title{Trim Table Directory for Inspections Tables}
\usage{
insp_trim_tables(
  table_dir = "V:/Compliance/Inspection Data for Publishing/Table/",
  table_pattern = "insp_table_.*html",
  min_tables = 1L
)
}
\arguments{
\item{table_dir}{The directory containing the most recent inspections
table(s)}

\item{table_pattern}{The inspections table naming pattern, in regex. This
prevents deletion of files which aren't inspection tables.}

\item{min_tables}{The minimum number of tables to keep in \code{table_dir}}
}
\value{
\code{table_dir} (invisibly)
}
\description{
\code{insp_trim_tables()} ensures that at least \code{min_tables} tables are
in the tables directory at a time, but that at most \code{min_tables} days worth
of tables are kept if there are more than \code{min_tables} tables. All tables
should be archived before removal. This function is for internal use in
\code{
\link[covidReport:insp_archive_table]{insp_archive_table()}
}.
}
\seealso{
\code{
  \link[covidReport:insp_archive_table]{insp_archive_table()}
  }, \code{\link[coviData:trim_backups]{trim_backups()}}
}
\keyword{internal}
