% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{inspections_cast_chr_date}
\alias{inspections_cast_chr_date}
\title{Cast Character Dates to a Standard Format}
\usage{
inspections_cast_chr_date(x)
}
\arguments{
\item{x}{A vector of dates}
}
\value{
The standardized date vector
}
\description{
\code{inspections_cast_chr_date()} converts dates in a variety of character
formats to the ISO 8601 standard. It is intended for use as the
\code{character_fun} in \code{\link[janitor:convert_to_date]{convert_to_date()}}.
}
\seealso{
The function that this is used in,
\code{\link[janitor:convert_to_date]{convert_to_date()}}, as well as the
higher-level function
\code{
  \link[covidReport:inspections_prep_data]{inspections_prep_data()}
  }
}
\keyword{internal}
