% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{inspections_cols_exist}
\alias{inspections_cols_exist}
\title{Check that the Expected Columns are Present in Inspections Data}
\usage{
inspections_cols_exist(.data)
}
\arguments{
\item{.data}{A data frame containing the inspections data}
}
\value{
The input data frame
}
\description{
\code{inspections_cols_exist()} is a wrapper around
\code{\link[coviData:cols_exist]{cols_exist()}} that checks for the existence
of columns expected in the inspections data. It is intended for use in
\code{\link[covidReport:inspections_load_data]{inspections_load_data()}}.
}
\seealso{
\code{
  \link[covidReport:inspections_load_data]{inspections_load_data()}
  }
}
\keyword{internal}
