% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{inspections_create_table}
\alias{inspections_create_table}
\title{Create HTML Table of COVID-19 Business Inspections Results}
\usage{
inspections_create_table(.data)
}
\arguments{
\item{.data}{The prepared inspections data}
}
\value{
A \code{datatables} object
}
\description{
\code{inspections_create_table()} takes prepared inspections data and creates an
HTML table for publishing on the web (using the \strong{DT} package). Data
should first be prepared using
\code{\link[covidReport:inspections_prep_data]{inspections_prep_data()}}.
}
\seealso{
Other functions in the inspections data pipeline (
\code{
  \link[covidReport:inspections_load_data]{inspections_load_data()}
  },
\code{
  \link[covidReport:inspections_prep_data]{inspections_prep_data()}
  },
\code{
  \link[covidReport:inspections_save_table]{inspections_save_table()}
  },
\code{
  \link[covidReport:inspections_archive_table]{inspections_archive_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:inspections_table_pipeline]{inspections_table_pipeline()}
  })
}
