% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{inspections_prep_data}
\alias{inspections_prep_data}
\title{Prepare Inspections Data for Displaying in HTML Table}
\usage{
inspections_prep_data(.data)
}
\arguments{
\item{.data}{A data frame containing the inspections data}
}
\value{
The processed data frame
}
\description{
\code{inspections_prep_data()} prepares data from the inspections team for display
as an HTML table. It selects, transforms, orders, and sorts the five columns
needed for display and outputs a visualization-ready dataset.
}
\details{
\code{inspections_prep_data()} chains together a number of operations on the input
data.

First, columns are selected as defined in the \strong{Details} of
\code{\link[covidReport:inspections_cols_exist]{inspections_cols_exist()}};
if multiple columns match this selection, they are
\code{\link[coviData:coalesce_across]{coalesced}}.

Second, columns are transformed according to type and assigned a standard
name. Visit date and closure date are assigned to \code{dt_visit} and \code{dt_closed},
respecively, and converted to the ISO 8601 standard (see
\code{\link[janitor:convert_to_date]{convert_to_date()}},
\code{
\link[covidReport:inspections_cast_chr_date]{inspections_cast_chr_date()}
}, and
\code{
\link[covidReport:inspections_replace_year_long]{
inspections_replace_year_long()
}}). Name and address columns are assigned to \code{nm_business} and
\code{addr_business}, respectively, and converted to character with
extra whitespace removed. The number of violations is assigned to
\code{n_violations} and converted to integer.

Third, rows missing a business name or number of violations (\code{nm_business}
or \code{n_violations}) are removed from the dataset. Entries without names are
not easily identifiable, and the primary purpose of the table is to display
violations.

Fourth, columns are reordered for ease of viewing, and unnecessary columns
(any not listed above) are dropped. The ordering is \code{n_violations},
\code{nm_business}, \code{addr_business}, \code{dt_visit}, and \code{dt_closed}.

Last, rows are arranged by column from left to right (first by
\code{n_violations}, then by \code{nm_business}, and so on). \code{n_violations},
\code{dt_visit}, and \code{dt_closed} are arranged in descending order (most violations
and most recent dates first); \code{nm_business} and \code{addr_business} are arranged
in ascending order (aka alphabetically).
}
\seealso{
Other functions in the inspections data pipeline (
\code{
  \link[covidReport:inspections_load_data]{inspections_load_data()}
  },
\code{
  \link[covidReport:inspections_create_table]{inspections_create_table()}
  },
\code{
  \link[covidReport:inspections_save_table]{inspections_save_table()}
  },
\code{
  \link[covidReport:inspections_archive_table]{inspections_archive_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:inspections_table_pipeline]{inspections_table_pipeline()}
  })
}
