% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{inspections_save_table}
\alias{inspections_save_table}
\title{Save an HTML Table Produced by \code{inspections_create_table()}}
\usage{
inspections_save_table(
  .table,
  path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/",
    paste0("inspections_table_", Sys.Date()), ext = "html"),
  self_contained = TRUE,
  force = FALSE
)
}
\arguments{
\item{.table}{A \code{datatables} object}

\item{path}{The location to save the table}

\item{force}{Should an existing file at \code{path} be overwritten if a
conflict occurs?}
}
\value{
The input \code{.table} (invisibly)
}
\description{
\code{inspections_save_table()} saves a \code{datatables} object to the specified
\code{path}. The defaults are intended for saving the inspections data table.
}
\seealso{
Other functions in the inspections data pipeline (
\code{
  \link[covidReport:inspections_load_data]{inspections_load_data()}
  },
\code{
  \link[covidReport:inspections_prep_data]{inspections_prep_data()}
  },
\code{
  \link[covidReport:inspections_create_table]{inspections_create_table()}
  },
\code{
  \link[covidReport:inspections_archive_table]{inspections_archive_table()}
  }) and the wrapper for the full workflow
(\code{
  \link[covidReport:inspections_table_pipeline]{inspections_table_pipeline()}
  })
}
