% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table.R
\name{inspections_table_pipeline}
\alias{inspections_table_pipeline}
\title{Run the Full Inspections Table Pipeline}
\usage{
inspections_table_pipeline(
  data_path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/",
    "Grand List of Inspections", ext = "xlsx"),
 
    table_path = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/",
    paste0("inspections_table_", Sys.Date()), ext = "html"),
 
    archive_dir = coviData::path_create("V:/Compliance/Inspection Data for Publishing/Table/Archive/"),
  force = FALSE
)
}
\arguments{
\item{data_path}{The path to the excel workbook containing inspections data}

\item{table_path}{The location to save the inspections table}

\item{archive_dir}{The path to the archive directory for inspections tables}

\item{force}{Should existing files at \code{table_path} or \code{archive_dir} be
overwritten if a conflict occurs?}
}
\value{
The inspections table as a \code{datatables} object (invisibly)
}
\description{
This is a simplified wrapper for the inspections table pipeline that makes
it easier to run. This takes inspections data from load to finished product
and cleans up after itself. See the component functions for details on the
pipeline.
}
\seealso{
The component functions of the inspections table pipeline:
\code{
  \link[covidReport:inspections_load_data]{inspections_load_data()}
  },
\code{
  \link[covidReport:inspections_prep_data]{inspections_prep_data()}
  },
\code{
  \link[covidReport:inspections_create_table]{inspections_create_table()}
  },
\code{
  \link[covidReport:inspections_save_table]{inspections_save_table()}
  },
\code{
  \link[covidReport:inspections_archive_table]{inspections_archive_table()}
  }
}
