% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_table.R
\name{inv_table}
\alias{inv_table}
\title{Tabluate Summary Investigation Numbers}
\usage{
inv_table(
  data = coviData::process_positive_people(date),
  nit_token = Sys.getenv("redcap_NIT_token"),
  prior_contacts = 45669L,
  date = NULL
)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process_positive_people]{process_positive_people()}}}

\item{nit_token}{API token/key for accessing the NIT REDcap project}

\item{prior_contacts}{Number of contacts recorded prior to REDcap
implementation}

\item{date}{The download date of \code{data}; defaults to most recent}
}
\description{
Tabluate Summary Investigation Numbers
}
