% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peds_race.R
\name{peds_table_race}
\alias{peds_table_race}
\title{Tabluate Pediatric Cases by Race}
\usage{
peds_table_race(data = pos(process_inv(read_inv(date))), date = NULL)
}
\arguments{
\item{data}{NBS case data, as returned by
\code{\link[coviData:read-nbs]{pos(process_inv())}}}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
A \code{flextable}
}
\description{
Tabluate Pediatric Cases by Race
}
