% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{ph_location_table}
\alias{ph_location_table}
\title{Create a Location for a Flextable into a Powerpoint}
\usage{
ph_location_table(
  flextable,
  pptx,
  layout,
  master = NULL,
  type = "tbl",
  pos_h = TRUE,
  pos_first = TRUE,
  halign = 0.5,
  valign = 0.5,
  newlabel = "",
  id = 1
)
}
\arguments{
\item{flextable}{The flextable to add}

\item{pptx}{The powerpoint object}

\item{layout}{The layout of the slide being added to}

\item{master}{The slide master to use, if any}

\item{type}{The placeholder type to use}

\item{pos_h}{If multiple of the same placeholder \code{type} exist, how should
the placeholder be chosen? \code{TRUE} means the horizontal axis will be
used; \code{FALSE} means vertical. See \code{pos_first} for more information.}

\item{pos_first}{If multiple of the same placeholder \code{type} exist, should
the first (or last) be chosen (along the axis given by \code{pos_h})?}

\item{halign}{Horizontal table alignment with placeholder; \code{0} aligns left
borders, \code{1} aligns right borders. Defaults to centered (\code{0.5}).}

\item{valign}{Vertical table alignment with placeholder; \code{0} aligns bottom
borders, \code{1} aligns top borders. Defaults to centered (\code{0.5}).}

\item{newlabel}{a label for the placeholder. See section details.}

\item{id}{index of the placeholder template. If two body placeholder, there can be
two different index: 1 and 2 for the first and second body placeholders defined
in the layout.}
}
\description{
Create a Location for a Flextable into a Powerpoint
}
\keyword{internal}
