% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vaccinations.R
\name{plot_vaccinations}
\alias{plot_vaccinations}
\title{Plot Vaccination Totals Figure}
\usage{
plot_vaccinations(
  date = NULL,
  n_first = NULL,
  n_second = NULL,
  n_goal = 656016,
  n_max = 937166,
  date_updated = NULL,
  resident_only = TRUE
)
}
\arguments{
\item{date}{Date of file to pull; defaults to most recent}

\item{n_first}{Number of first doses given}

\item{n_second}{Number of second doses given}

\item{n_goal}{Goal for number of people vaccinated}

\item{n_max}{Maximum number of people (population size)}

\item{date_updated}{The date to use for the figure subtitle; defaults to last
updated date in the file}

\item{resident_only}{Should the figure only count Shelby County residents?}
}
\value{
A \code{ggplot} object
}
\description{
Plot Vaccination Totals Figure
}
