% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_age_2019.R
\docType{data}
\name{pop_2019}
\alias{pop_2019}
\title{Shelby County Age Demographics - ACS 2019 1-Year Estimates}
\format{
A \code{tibble} with 1376 rows and 5 variables:
\describe{
\item{age}{age in years}
\item{sex}{Male or Female}
\item{race}{
Black/African American,
White,
American Indian/Alaska Native,
Asian/Pacific Islander
}
\item{population}{number of people in group}
}
}
\source{
Vintage 2019 NCHS Bridged-Race Postcensal Population Estimates
}
\usage{
pop_2019
}
\description{
Shelby County Age Demographics - ACS 2019 1-Year Estimates
}
\keyword{datasets}
