% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_age_2019.R
\docType{data}
\name{pop_age_2019}
\alias{pop_age_2019}
\title{Shelby County Age Demographics - ACS 2019 1-Year Estimates}
\format{
A \code{tibble} with 8 rows and 2 variables:
\describe{
\item{age_grp}{age group, in years (closed on left, open on right)}
\item{n}{Number of people in age group}
}
}
\source{
US ACS Census Estimates 2019
}
\usage{
pop_age_2019
}
\description{
Shelby County Age Demographics - ACS 2019 1-Year Estimates
}
\keyword{datasets}
