% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_table.R
\name{prod_calc}
\alias{prod_calc}
\title{Calculate Productivity Summary Statistics}
\usage{
prod_calc(
  assigned = read_prod(sheet = "assign_dt_n"),
  nit_token = Sys.getenv("redcap_NIT_token"),
  days = 14,
  lag = 6
)
}
\arguments{
\item{assigned}{Assigned case data from \code{read_prod()}}

\item{nit_token}{API token to access NIT RedCAP project}

\item{days}{Number of days to summarize over}

\item{lag}{Number of days back to end the calculation}
}
\value{
A \code{tibble}
}
\description{
Calculate Productivity Summary Statistics
}
