% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_table.R
\name{prod_table}
\alias{prod_table}
\title{Productivity Summary Table}
\usage{
prod_table(
  assigned = read_prod(sheet = "assign_dt_n"),
  nit_token = Sys.getenv("redcap_NIT_token"),
  days = 14,
  lag = 6
)
}
\arguments{
\item{assigned}{Assigned case data from \code{read_prod()}}

\item{nit_token}{API token to access NIT RedCAP project}

\item{days}{Number of days to summarize over}

\item{lag}{Number of days back to end the calculation}
}
\value{
A \code{flextable}
}
\description{
Productivity Summary Table
}
