% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_table.R
\name{read_prod}
\alias{read_prod}
\title{Read Productivity Data}
\usage{
read_prod(
  sheet = c("assign_dt_n", "assign_dt_pct_assign", "assign_dt_pct_complete",
    "interview_dt"),
  file = "productivity_report.xlsx",
  dir = "V:/Productivity/Daily Report for Administration"
)
}
\arguments{
\item{sheet}{Excel sheet to read}

\item{file}{Name of productivity file - must be xlsx or xls}

\item{dir}{Directory location of productivity file}
}
\value{
A \code{tibble}
}
\description{
Read Productivity Data
}
