% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_table.R
\name{result_table}
\alias{result_table}
\title{Create Tables for Test Result}
\usage{
result_table(date = NULL, by = c("person", "test"))
}
\arguments{
\item{date}{The download date of the data to use; defaults to the most recent
data}

\item{by}{Units to report- either \code{"person"} or \code{"test"}}
}
\value{
A \code{gt_tbl} object
}
\description{
Create Tables for Test Result
}
