% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_daily.R
\name{rpt_daily_mail}
\alias{rpt_daily_mail}
\title{Send Daily COVID-19 Status Summary via Outlook Email}
\usage{
rpt_daily_mail(
  date = NULL,
  to = c("Jesse.Smith@shelbycountytn.gov", "Chaitra.Subramanya@shelbycountytn.gov",
    "Allison.Plaxco@shelbycountytn.gov")
)
}
\arguments{
\item{date}{The date for which to run the report; defaults to most recent}

\item{to}{A \code{character} vector of recipient email addresses}
}
\description{
Send Daily COVID-19 Status Summary via Outlook Email
}
