% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_daily.R
\name{rpt_daily_pptx}
\alias{rpt_daily_pptx}
\title{Create Powerpoint for Daily Status Report}
\usage{
rpt_daily_pptx(
  date = NULL,
 
    dir = coviData::path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Status Report",
    "automated")
)
}
\arguments{
\item{date}{The date for which to run the report}

\item{dir}{The directory to save the report; set to \code{NULL} to return without
saving}
}
\value{
A \code{pptx} object
}
\description{
Create Powerpoint for Daily Status Report
}
