% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a \code{ggplot} Object}
\usage{
save_plot(gg_obj, path, width = 16, height = 9, force = FALSE, ...)
}
\arguments{
\item{gg_obj}{A \code{ggplot} object to save}

\item{path}{The location for saving}

\item{width}{The width of the saved figure, in inches by default}

\item{height}{The height of the saved figure, in inches by default}

\item{force}{Should existing files be overwritten?}

\item{...}{Additional arguments to pass to
\code{\link[ggplot2:ggsave]{ggsave()}}}
}
\value{
The \code{gg_obj}, invisibly
}
\description{
\code{save_plot()} saves a \code{ggplot} object to the specified path. It is a wrapper
around \code{\link[ggplot2:ggsave]{ggsave()}} that takes the plot as the
first argument and provides alternate defaults for use in SCHD slides. Unlike
\code{ggsave()} the \code{ggplot} object must always be supplied explicitly, and it
will not overwrite existing files by default.
}
