% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_breaks.R
\name{scale_breaks}
\alias{scale_breaks}
\title{Create breaks for scales}
\usage{
scale_breaks(lower, upper, lim_zero = TRUE)
}
\arguments{
\item{min}{Minimum value in scale range}

\item{max}{Maximum value in scale range}
}
\value{
A numeric sequence of appropriately spaced break points
}
\description{
Create breaks for scales
}
\keyword{internal}
