% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by.R
\name{scale_by}
\alias{scale_by}
\title{Calculate Scale Guides for Plots}
\usage{
scale_by(range)
}
\arguments{
\item{range}{Either a scalar numeric indicating the range of the scale,
or a length 2 numeric indicating the min and max of the scale}
}
\value{
A \code{double}
}
\description{
Calculate Scale Guides for Plots
}
\keyword{internal}
