% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelby_poly.R
\docType{data}
\name{shelby_poly}
\alias{shelby_poly}
\title{Shelby County Polygon}
\format{
A \code{tibble} with 4420 rows and 2 columns:
\describe{
\item{x}{x-coordinate, scaled to 1}
\item{y}{y-coordinate, scaled to 1}
}
}
\source{
urbnmpr
}
\usage{
shelby_poly
}
\description{
Shelby County Polygon
}
\keyword{datasets}
