% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{skip_reports}
\alias{skip_reports}
\title{Skip Reporting Function Tests}
\usage{
skip_reports(skip = NULL)
}
\arguments{
\item{skip}{Should the reporting tests be skipped? If \code{SKIP_REPORTS} is unset
during testing, \code{NULL} defaults to \code{TRUE}.}
}
\value{
\code{NULL}
}
\description{
Reporting functions are time-consuming to run; \code{skip_reports()} sets an
environment variable that allows the user to skip these tests. Component
functions should be tested separately.
}
\details{
When \code{skip = NULL}, this is used as a skipping function within a test. When
a value is supplied, it sets the \code{SKIP_REPORTS} environment variable that
controls behavior during testing.
}
\keyword{internal}
