% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_table_total.R
\name{test_calc_total}
\alias{test_calc_total}
\title{Calculate PCR Test Totals}
\usage{
test_calc_total(data = process_pcr(read_pcr(date)), date = NULL)
}
\arguments{
\item{data}{PCR test data, read from an NBS snapshot file}

\item{date}{The download date of the data to read; defaults to most recent}
}
\value{
A \code{tibble}
}
\description{
Calculate PCR Test Totals
}
\keyword{internal}
