% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_map_rate.R
\name{test_map_rate}
\alias{test_map_rate}
\title{Plot Test Rates by ZIP Code for a Given Period}
\usage{
test_map_rate(
  data = process_pcr(read_pcr(date = date)),
  days = 14L,
  lag = 5L,
  date = NULL
)
}
\arguments{
\item{data}{Prepped case data}

\item{days}{Number of days to consider}

\item{lag}{Reporting lag}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
A \code{ggplot} object
}
\description{
Plot Test Rates by ZIP Code for a Given Period
}
