% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plot_positivity.R
\name{test_plot_positivity}
\alias{test_plot_positivity}
\title{Plot 7-Day Rolling Test Positivity Rate}
\usage{
test_plot_positivity(
  data = process_pcr(read_pcr(date)),
  date = NULL,
  delay = 5L
)
}
\arguments{
\item{data}{PCR test data, read from an NBS snapshot file}

\item{date}{The download date of the data to read; defaults to most recent}

\item{delay}{\code{integer}. The reporting delay of the data; test dates within
\code{delay} days of \code{date} will be omitted}
}
\value{
A \code{gt_tbl}
}
\description{
Plot 7-Day Rolling Test Positivity Rate
}
