% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_table_total.R
\name{test_table_total}
\alias{test_table_total}
\title{Tabulate PCR Test Totals}
\usage{
test_table_total(
  data = coviData::read_file_delim(coviData::path_pcr(date)),
  date = NULL
)
}
\arguments{
\item{data}{PCR test data, read from an NBS snapshot file}

\item{date}{The download date of the data to read; defaults to most recent}
}
\value{
A \code{flextable}
}
\description{
Tabulate PCR Test Totals
}
