% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_map_pct.R
\name{vac_map_pct}
\alias{vac_map_pct}
\title{Plot \% Vaccinated by ZIP Code}
\usage{
vac_map_pct(
  data = coviData::vac_prep(coviData::read_vac(date = date)),
  date = NULL,
  zip_path = coviData::path_create("O:/EPI/Shapefiles 07.2014/MergedZips",
    "AllZipswithMergedZips_clippedbySCBoundary.shp")
)
}
\arguments{
\item{data}{Prepped vaccination data from TennIIS}

\item{date}{The download date of the data; defaults to most recent}

\item{zip_path}{Path to Shelby ZIP shapefile}
}
\value{
A \code{ggplot} object
}
\description{
Plot \% Vaccinated by ZIP Code
}
