% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_plot_goal.R
\name{vac_plot_goal}
\alias{vac_plot_goal}
\title{Plot Vaccination Totals Figure}
\usage{
vac_plot_goal(
  data = coviData::vac_prep(coviData::read_vac(date = date)),
  n_goal = 0.7 * n_max,
  n_max = 937166,
  date = NULL
)
}
\arguments{
\item{data}{Vaccination date from \code{\link[coviData:vac_prep]{vac_prep()}}}

\item{n_goal}{Goal for number of people vaccinated}

\item{n_max}{Maximum number of people (population size)}

\item{date}{Date of file to pull; defaults to most recent}

\item{n_vaccinated}{Number of people vaccinated}

\item{n_first}{Deprecated (use \code{n_vaccinated}). Number of first doses given}

\item{n_second}{Deprecate (no longer used). Number of second doses given}

\item{date_updated}{The date to use for the figure subtitle; defaults to last
updated date in the file}

\item{resident_only}{Should the figure only count Shelby County residents?}
}
\value{
A \code{ggplot} object
}
\description{
Plot Vaccination Totals Figure
}
