% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{vac_residents}
\alias{vac_residents}
\title{Filter Vaccination Data to Possible Shelby County Residents}
\usage{
vac_residents(
  data = coviData::vac_prep(coviData::read_vac(date = date)),
  date = NULL
)
}
\arguments{
\item{data}{Vaccination data with \code{logical} \code{resident} column}

\item{date}{The download date of the data; defaults to most recent}
}
\value{
\code{data} filtered to potential residents
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated. Please use the
\code{\link[coviData:vac_filter_residents]{vac_filter_residents()}} function,
which is now the backend for this function.
}
\keyword{internal}
