% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_table_doses.R
\name{vac_table_doses}
\alias{vac_table_doses}
\title{Create Table of Vaccination Dose Totals}
\usage{
vac_table_doses(
  data = coviData::vac_prep(coviData::vac_load(date = date)),
  date = NULL
)
}
\arguments{
\item{data}{TennIIS data}

\item{date}{Download date}
}
\value{
\code{gt_tbl}
}
\description{
Create Table of Vaccination Dose Totals
}
