% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_table_recent.R
\name{vac_table_recent}
\alias{vac_table_recent}
\title{Create Table of Recent and Total Vaccine Doses}
\usage{
vac_table_recent(
  data = coviData::vac_prep(coviData::read_vac(date = date)),
  date = NULL
)
}
\arguments{
\item{data}{TennIIS vaccination data, as output from
\code{\link[coviData:vac_prep]{vac_prep()}}}

\item{date}{The download date of the data to use; defaults to the most recent
file}
}
\value{
A \code{gt_tbl}
}
\description{
Create Table of Recent and Total Vaccine Doses
}
