% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_table_totals.R
\name{vac_table_totals}
\alias{vac_table_totals}
\title{Create a Table of Vaccination Totals}
\usage{
vac_table_totals(
  data = coviData::vac_prep(coviData::vac_load(date = date)),
  date = NULL
)
}
\arguments{
\item{data}{TennIIS vaccination data, as output by
\code{\link[coviData:vac_prep]{vac_prep()}}}

\item{date}{The download date of the data to use; defaults to most recent
file}
}
\value{
A \code{gt_tbl}
}
\description{
Create a Table of Vaccination Totals
}
