% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspections_table2.R
\name{violation_style}
\alias{violation_style}
\title{Inspections Violations Color Helper}
\usage{
violation_style(value)
}
\arguments{
\item{value}{A scalar or vector numeric. This function \emph{can} accept vectors;
however, \code{\link[reactable:colDef]{colDef()}} passes scalars.}
}
\value{
A hex color code
}
\description{
Map number of violations to a color (green, orange, or red). Invalid values
default to black (\code{"#000000"}).
}
