% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_shape.R
\docType{data}
\name{zip_shape}
\alias{zip_shape}
\title{Merge Shelby County ZIP Code Geometries}
\format{
\describe{
\item{\code{zip}}{\code{chr}: Merged ZIP code labels}
\item{\code{pop_2019}}{\code{int}: ACNS 2019 5-year population estimates}
\item{\code{geometry}}{\code{MULTIPOLYGON}: shape geometries}
}
}
\source{
Shelby County Health Department
}
\usage{
zip_shape
}
\description{
Merge Shelby County ZIP Code Geometries
}
\keyword{datasets}
