% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_plot_cumulative.R
\name{case_plot_cumulative}
\alias{case_plot_cumulative}
\title{Plot Cumulative Case Time Series}
\usage{
case_plot_cumulative(
  data = pos(process_inv(read_inv(date = date))),
  date = NULL
)
}
\arguments{
\item{data}{Case data (by person) from NBS, as output by
\code{\link[coviData:process-nbs]{pos(process_inv())}}}

\item{date}{The download date of \code{data}; the default is the most recent}
}
\value{
A \code{ggplot} object
}
\description{
Plot Cumulative Case Time Series
}
