% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_plot_daily.R
\name{case_plot_daily}
\alias{case_plot_daily}
\title{Plot Daily New Cases by Specimen Collection Date}
\usage{
case_plot_daily(
  data = coviData::process_positive_people(date = date),
  date = NULL,
  delay = NULL
)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process_positive_people]{process_positive_people()}}}

\item{date}{The report date of the data; defaults to the most recent date}

\item{delay}{Number of days to ignore 7 day average (due to incomplete data);
default is calculated using
\code{\link[covidModel:estimate_delay]{estimate_delay()}}}
}
\value{
A \code{ggplot} object
}
\description{
Plot Daily New Cases by Specimen Collection Date
}
