% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_table_total.R
\name{death_table_total}
\alias{death_table_total}
\title{Tabulated Death and Survival Outcomes from COVID-19}
\usage{
death_table_total(data = pos(process_inv(read_inv(date = date))), date = NULL)
}
\arguments{
\item{data}{Case data, as output by
\code{\link[coviData:process-nbs]{pos(process_inv())}}}

\item{date}{The download data of the data; defaults to most recent}
}
\value{
A \code{flextable}
}
\description{
Tabulated Death and Survival Outcomes from COVID-19
}
