% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_plot_.R
\name{demog_plot_}
\alias{demog_plot_}
\title{Plot Demographic Summary}
\usage{
demog_plot_(
  data,
  unit,
  grp = c("age", "sex", "race", "ethnicity"),
  color = "midnightblue",
  vjust = c("top", "bottom"),
  date = NULL
)
}
\arguments{
\item{data}{Data from an \verb{active_calc_*()} function}

\item{grp}{The grouping variable}
}
\value{
A \code{ggplot}
}
\description{
Plot Demographic Summary
}
\keyword{internal}
