% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_covid_table.R
\name{fmt_covid_table}
\alias{fmt_covid_table}
\alias{fmt_covid_table.flextable}
\alias{fmt_covid_table.gt_tbl}
\title{Standardize Table Styling for COVID-19 Reports}
\usage{
fmt_covid_table(
  table,
  total = FALSE,
  align_label = c("left", "center", "right")
)

\method{fmt_covid_table}{flextable}(
  table,
  total = FALSE,
  align_label = c("left", "center", "right")
)

\method{fmt_covid_table}{gt_tbl}(
  table,
  total = FALSE,
  align_label = c("left", "center", "right")
)
}
\arguments{
\item{table}{A \code{flextable} or \code{gt_tbl} object}

\item{total}{Does the table contain a "Total" row at the bottom?}

\item{align_label}{Alignment for left-most column (labels, by convention)}
}
\value{
A \code{flextable} or \code{gt_tbl} with modified styling
}
\description{
\code{fmt_covid_table()} standardizes styling for \code{flextable} and \code{gt_tbl} objects
}
\keyword{internal}
