% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_vac_age}
\alias{plot_vac_age}
\title{Plot Vaccinations by Age}
\usage{
plot_vac_age(
  .data = coviData::vac_prep(coviData::read_vac()),
  pct = TRUE,
  by_pop = TRUE,
  incl_under_15 = FALSE,
  incl_under_12 = incl_under_15
)
}
\arguments{
\item{.data}{Vaccination data, as created by
\code{\link[coviData:vac_prep]{vac_prep()}}}

\item{pct}{Deprecated.}

\item{by_pop}{Should the resulting graphic be calculated using age group
population statistics (\code{TRUE}, the default) or not (\code{FALSE})?}

\item{incl_under_15}{Deprecated.}

\item{incl_under_12}{Should the age group \code{"0-11"} be included? The default
is \code{FALSE}.}
}
\value{
A \code{ggplot} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated. Please use the
\code{\link[covidReport:vac_plot_age]{vac_plot_age()}} function, which is now
the backend for this function.
}
\keyword{internal}
