% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_vaccinations}
\alias{plot_vaccinations}
\title{Plot Vaccination Goal}
\usage{
plot_vaccinations(
  data = coviData::vac_prep(coviData::read_vac(date = date)),
  date = NULL,
  n_vaccinated = NULL,
  n_first = NULL,
  n_second = NULL,
  n_goal = 7e+05,
  n_max = 937166,
  date_updated = NULL,
  resident_only = TRUE
)
}
\arguments{
\item{data}{Vaccination date from \code{\link[coviData:vac_prep]{vac_prep()}}}

\item{date}{Date of file to pull; defaults to most recent}

\item{n_goal}{Goal for number of people vaccinated}

\item{n_max}{Maximum number of people (population size)}
}
\value{
A \code{ggplot} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated. Please use the
\code{\link[covidReport:vac_plot_goal]{vac_plot_goal()}} function, which is
now the backend for this function.
}
\keyword{internal}
