% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_daily.R
\name{rpt_daily_mail}
\alias{rpt_daily_mail}
\title{Send Daily COVID-19 Status Summary via Outlook Email}
\usage{
rpt_daily_mail(
  date = NULL,
  to = c("Jesse.Smith@shelbycountytn.gov", "Chaitra.Subramanya@shelbycountytn.gov",
    "Allison.Plaxco@shelbycountytn.gov"),
 
    dir_pptx = coviData::path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Status Report",
    "automated"),
  demog = rlang::is_true(weekdays(date_inv(date)) == "Tuesday"),
  inv = process_inv(read_inv(date)),
  pcr = process_pcr(read_pcr(date), inv = inv)
)
}
\arguments{
\item{date}{The date for which to run the report; defaults to most recent}

\item{to}{A \code{character} vector of recipient email addresses}

\item{dir_pptx}{Directory containing daily status report powerpoint files}

\item{demog}{Should the demographics report be attached? The default attaches
on Tuesdays only.}

\item{inv}{Investigation data from
\code{\link[coviData:process-nbs]{process_inv()}}}

\item{pcr}{PCR test data from
\code{\link[coviData:process-nbs]{process_pcr()}}}
}
\description{
Send Daily COVID-19 Status Summary via Outlook Email
}
