% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_daily.R
\name{rpt_daily_pptx}
\alias{rpt_daily_pptx}
\title{Create Powerpoint for Daily Status Report}
\usage{
rpt_daily_pptx(
  date = NULL,
 
    dir = coviData::path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Status Report",
    "automated"),
  inv = process_inv(read_inv(date)),
  pcr = process_pcr(read_pcr(date), inv = inv)
)
}
\arguments{
\item{date}{The date for which to run the report; defaults to most recent}

\item{dir}{The directory to save the report; set to \code{NULL} to return without
saving}

\item{inv}{Investigation data from
\code{\link[coviData:process-nbs]{process_inv()}}}

\item{pcr}{PCR test data from
\code{\link[coviData:process-nbs]{process_pcr()}}}
}
\value{
An \code{rpptx} object
}
\description{
Create Powerpoint for Daily Status Report
}
