% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_reports.R
\name{skip_reports}
\alias{skip_reports}
\title{Skip Reporting Function Tests}
\usage{
skip_reports(skip = NULL)
}
\arguments{
\item{skip}{Should the reporting tests be skipped? If
\code{getOption(SKIP_REPORTS)} is unset during testing, \code{NULL} defaults to
\code{TRUE}.}
}
\value{
Either the previous value of \code{SKIP_REPORTS} (if setting the option)
or a \code{skip} condition
}
\description{
Reporting functions are time-consuming to run; \code{skip_reports()} sets an
option that allows the user to skip these tests. Component functions should
be tested separately.
}
\details{
When \code{skip = NULL}, this is used as a skipping function within a test. When
a value is supplied, it sets the \code{SKIP_REPORTS} option that controls
behavior during testing.
}
\keyword{internal}
