% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_count.R
\name{vac_count}
\alias{vac_count}
\title{Count Vaccinations by Dose}
\usage{
vac_count(
  .data = coviData::vac_prep(coviData::vac_load(date = date)),
  resident_only = TRUE,
  date = NULL
)
}
\arguments{
\item{.data}{Vaccination data; default loads and parses the latest file}

\item{resident_only}{Should counts include Shelby County residents only? If
so, persons of unknown residency will still be included.}

\item{date}{The date of the vaccination data to load; defaults to latest
data}
}
\value{
A \code{tibble} with columns \code{dose_count} and \code{n}
}
\description{
Count Vaccinations by Dose
}
