% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_date.R
\name{vac_date}
\alias{vac_date}
\title{Get Latest Date Contained in Vaccination File}
\usage{
vac_date(date = NULL)
}
\arguments{
\item{date}{The date of the file to use; defaults to most recent. If you
know this, you don't actually need this function; it's here for
compatiblity reasons only.}
}
\description{
Get Latest Date Contained in Vaccination File
}
