% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_date.R
\name{vac_date}
\alias{vac_date}
\title{Get Latest Date Contained in Vaccination File}
\usage{
vac_date(date = NULL, resident_only = TRUE)
}
\arguments{
\item{date}{The date of the file to use; defaults to most recent}

\item{resident_only}{Should non-residents be filtered from the data?}
}
\value{
A \code{Date}
}
\description{
Get Latest Date Contained in Vaccination File
}
