% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_plot_age.R
\name{vac_plot_age}
\alias{vac_plot_age}
\title{Plot Vaccinations by Age Group}
\usage{
vac_plot_age(
  .data = vac_prep(read_vac(date)),
  date = NULL,
  by_pop = TRUE,
  incl_under_12 = FALSE
)
}
\arguments{
\item{.data}{Vaccination data, as created by
\code{\link[coviData:vac_prep]{vac_prep()}}}

\item{date}{Date of vaccination data}

\item{by_pop}{Should the resulting graphic be calculated using age group
population statistics (\code{TRUE}, the default) or not (\code{FALSE})?}

\item{incl_under_12}{Should the age group \code{"0-11"} be included? The default
is \code{FALSE}.}
}
\value{
A \code{ggplot} object
}
\description{
\code{vac_plot_age()} plots vaccinations by age group in a bar chart.
}
\details{
When \code{by_pop = TRUE},
the resulting figure displays vaccinations with reference to the population
of each age group; when \code{by_pop = FALSE}, the figure does not display
population-related values.
}
