% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac_plot_goal.R
\name{vac_plot_goal}
\alias{vac_plot_goal}
\title{Plot Vaccination Totals Figure}
\usage{
vac_plot_goal(
  data = coviData::vac_prep(coviData::read_vac(date = date)),
  date = NULL,
  n_goal = 0.7 * n_max,
  n_max = 937166
)
}
\arguments{
\item{data}{Vaccination date from \code{\link[coviData:vac_prep]{vac_prep()}}}

\item{date}{Date of file to pull; defaults to most recent}

\item{n_goal}{Goal for number of people vaccinated}

\item{n_max}{Maximum number of people (population size)}
}
\value{
A \code{ggplot} object
}
\description{
Plot Vaccination Totals Figure
}
